requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
slowapi
python-decouple
pyjwt
passlib
bcrypt
python-multipart
email-validator
colorlog
chacc-dependency-manager